extract removeEmptyDir to a new component
