+++ scopes/toolbox/fs/remove-empty-dir/index.ts
@@ -1,0 +1,1 @@
+export { removeEmptyDir } from './remove-empty-dir';
