+++ scopes/toolbox/fs/remove-empty-dir/remove-empty-dir.ts
